refactor: 식당 유틸 위치 이동
