+++ src/pages/Cafeteria/PCCafeteriaPage/components/PCDiningBlocks/index.tsx
@@ -1,15 +1,14 @@
 /* eslint-disable no-param-reassign */
-import { Dining, DiningType } from 'static/cafeteria';
+import { Dining, DiningType, DINING_TYPE_MAP } from 'static/cafeteria';
 import { useEffect, useRef } from 'react';
 import useLogger from 'utils/hooks/analytics/useLogger';
 import { useDatePicker } from 'pages/Cafeteria/hooks/useDatePicker';
 import useDinings from 'pages/Cafeteria/hooks/useDinings';
 import DetailModal from 'pages/Cafeteria/PCCafeteriaPage/components/DetailModal';
 import PCMealImage from 'pages/Cafeteria/PCCafeteriaPage/components/PCMealImage';
-import { DINING_TYPE_MAP } from 'static/cafeteria';
-import { filterDinings } from 'utils/ts/cafeteria';
 import useModalPortal from 'utils/hooks/layout/useModalPortal';
 import { Portal } from 'components/Modal/PortalProvider';
+import { filterDinings } from 'pages/Cafeteria/utils/filter';
 import styles from './PCDiningBlocks.module.scss';
 
 interface PCDiningBlocksProps {
